refactor(app): Import from src folder
