+++ app/(auth-pages)/forgot-password/page.tsx
@@ -1,8 +1,8 @@
 import { forgotPasswordAction } from "@/app/actions";
-import { FormMessage, Message } from "@/components/form-message";
-import { SubmitButton } from "@/components/submit-button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+import { FormMessage, Message } from "@/src/components/form-message";
+import { SubmitButton } from "@/src/components/submit-button";
+import { Input } from "@/src/components/ui/input";
+import { Label } from "@/src/components/ui/label";
 import Link from "next/link";
 import { SmtpMessage } from "../smtp-message";
