+++ app/protected/page.tsx
@@ -1,4 +1,4 @@
-import FetchDataSteps from "@/components/tutorial/fetch-data-steps";
+import FetchDataSteps from "@/src/components/tutorial/fetch-data-steps";
 import { createClient } from "@/utils/supabase/server";
 import { InfoIcon } from "lucide-react";
 import { redirect } from "next/navigation";
